+++ src-tauri/src/analytics/algo_controller.rs
