+++ src-tauri/src/analytics/algo_store.rs
