+++ src-tauri/src/analytics/algorithms/articulation_point.rs
