+++ src-tauri/src/analytics/algorithms/globalmincut.rs
