+++ src-tauri/src/analytics/algorithms/stoer_wagner.rs
