+++ src-tauri/src/analytics/algos_config.rs
