+++ src-tauri/src/analytics/mod.rs
